+++ tests/test_hl7.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from hl7 import Message, Segment, Field
+from __future__ import unicode_literals
+from hl7 import Message, Segment, Field, Repetition, Component
 
 import hl7
 import six
@@ -8,13 +9,35 @@ import unittest
 
 ## Sample message from HL7 Normative Edition
 ## http://healthinfo.med.dal.ca/hl7intro/CDA_R2_normativewebedition/help/v3guide/v3guide.htm#v3gexamples
-sample_hl7 = u'\r'.join([
+sample_hl7 = '\r'.join([
     'MSH|^~\&|GHH LAB|ELAB-3|GHH OE|BLDG4|200202150930||ORU^R01|CNTRL-3456|P|2.4',
     'PID|||555-44-4444||EVERYWOMAN^EVE^E^^^^L|JONES|196203520|F|||153 FERNWOOD DR.^^STATESVILLE^OH^35292||(206)3345232|(206)752-121||||AC555444444||67-A4335^OH^20030520',
     'OBR|1|845439^GHH OE|1045813^GHH LAB|1554-5^GLUCOSE|||200202150730||||||||555-55-5555^PRIMARY^PATRICIA P^^^^MD^^LEVEL SEVEN HEALTHCARE, INC.|||||||||F||||||444-44-4444^HIPPOCRATES^HOWARD H^^^^MD',
     'OBX|1|SN|1554-5^GLUCOSE^POST 12H CFST:MCNC:PT:SER/PLAS:QN||^182|mg/dl|70_105|H|||F',
     'OBX|2|FN|1553-5^GLUCOSE^POST 12H CFST:MCNC:PT:SER/PLAS:QN||^182|mg/dl|70_105|H|||F\r'
 ])
+
+# Example from: http://wiki.medical-objects.com.au/index.php/Hl7v2_parsing
+rep_sample_hl7 = '\r'.join([
+    'MSH|^~\&|GHH LAB|ELAB-3|GHH OE|BLDG4|200202150930||ORU^R01|CNTRL-3456|P|2.4',
+    'PID|Field1|Component1^Component2|Component1^Sub-Component1&Sub-Component2^Component3|Repeat1~Repeat2',
+    ''
+    ])
+
+# Source: http://www.health.vic.gov.au/hdss/vinah/2006-07/appendix-a-sample-messages.pdf
+sample_file = '\r'.join([
+    'FHS|^~\&||ABCHS||AUSDHSV|20070101123401|||abchs20070101123401.hl7|',
+    'BHS|^~\&||ABCHS||AUSDHSV|20070101123401||||abchs20070101123401-1',
+    'MSH|^~\&||ABCHS||AUSDHSV|20070101112951||ADT^A04^ADT_A01|12334456778890|P|2.5|||NE|NE|AU|ASCII',
+    'EVN|A04|20060705000000',
+    'PID|1||0000112234^^^100^A||XXXXXXXXXX^^^^^^S||10131113|1||4|^^RICHMOND^^3121||||1201||||||||1100|||||||||AAA',
+    'PD1||2',
+    'NK1|1||1||||||||||||||||||2',
+    'PV1|1|O||||^^^^^1',
+    'BTS|1',
+    'FTS|1',
+    ''
+    ])
 
 
 class ParseTest(unittest.TestCase):
@@ -26,8 +49,18 @@ class ParseTest(unittest.TestCase):
         self.assertEqual(msg[3][0][0], 'OBX')
         self.assertEqual(
             msg[3][3],
-            ['1554-5', 'GLUCOSE', 'POST 12H CFST:MCNC:PT:SER/PLAS:QN']
+            [[['1554-5'], ['GLUCOSE'], ['POST 12H CFST:MCNC:PT:SER/PLAS:QN']]]
         )
+        ## Make sure MSH-1 and MSH-2 are valid
+        self.assertEqual(msg[0][1][0], '|')
+        self.assertTrue(isinstance(msg[0][1], hl7.Field))
+        self.assertEqual(msg[0][2][0], '^~\&')
+        self.assertTrue(isinstance(msg[0][2], hl7.Field))
+        ## MSH-9 is the message type
+        self.assertEqual(msg[0][9], [[['ORU'], ['R01']]])
+        ## Do it twice to make sure text coercion is idempotent
+        self.assertEqual(six.text_type(msg), sample_hl7.strip())
+        self.assertEqual(six.text_type(msg), sample_hl7.strip())
 
     def test_bytestring_converted_to_unicode(self):
         msg = hl7.parse(six.text_type(sample_hl7))
@@ -39,7 +72,7 @@ class ParseTest(unittest.TestCase):
         # \x96 - valid cp1252, not valid utf8
         # it is the responsibility of the caller to convert to unicode
         msg = hl7.parse(b'MSH|^~\&|GHH LAB|ELAB\x963', encoding='cp1252')
-        self.assertEqual(msg[0][3][0], u'ELAB\u20133')
+        self.assertEqual(msg[0][4][0], 'ELAB\u20133')
 
     def test_non_ascii_bytestring_no_encoding(self):
         # \x96 - valid cp1252, not valid utf8
@@ -61,7 +94,67 @@ class ParseTest(unittest.TestCase):
 
         self.assertEqual(six.text_type(msg), nonstd)
         self.assertEqual(len(msg), 2)
-        self.assertEqual(msg[1][5], ['EVERYWOMAN', 'EVE', 'E', '', '', 'L'])
+        self.assertEqual(msg[1][5], [[['EVERYWOMAN'], ['EVE'], ['E'], [''], [''], ['L']]])
+
+    def test_repetition(self):
+        msg = hl7.parse(rep_sample_hl7)
+        self.assertEqual(msg[1][4], [['Repeat1'], ['Repeat2']])
+        self.assertEqual(type(msg[1][4]), Field)
+        self.assertEqual(type(msg[1][4][0]), Repetition)
+        self.assertEqual(type(msg[1][4][1]), Repetition)
+        self.assertEqual(six.text_type(msg[1][4][0][0]), 'Repeat1')
+        self.assertTrue(isinstance(msg[1][4][0][0], six.text_type))
+        self.assertEqual(six.text_type(msg[1][4][1][0]), 'Repeat2')
+        self.assertTrue(isinstance(msg[1][4][1][0], six.text_type))
+
+    def test_subcomponent(self):
+        msg = hl7.parse(rep_sample_hl7)
+        self.assertEqual(
+            msg[1][3],
+            [[['Component1'], ['Sub-Component1', 'Sub-Component2'], ['Component3']]]
+        )
+
+    def test_unescape(self):
+        msg = hl7.parse(rep_sample_hl7)
+
+        # Replace Separators
+        self.assertEqual(msg.unescape('\\E\\'), '\\')
+        self.assertEqual(msg.unescape('\\F\\'), '|')
+        self.assertEqual(msg.unescape('\\S\\'), '^')
+        self.assertEqual(msg.unescape('\\T\\'), '&')
+        self.assertEqual(msg.unescape('\\R\\'), '~')
+
+        # Replace Highlighting
+        self.assertEqual(msg.unescape('\\H\\text\\N\\'), '_text_')
+
+        # Application Overrides
+        self.assertEqual(msg.unescape('\\H\\text\\N\\', {'H': '*', 'N': '*'}), '*text*')
+
+        # Hex Codes
+        self.assertEqual(msg.unescape('\\X20202020\\'), '    ')
+
+    def test_escape(self):
+        msg = hl7.parse(rep_sample_hl7)
+
+        self.assertEqual(msg.escape('\\'), '\\E\\')
+        self.assertEqual(msg.escape('|'), '\\F\\')
+        self.assertEqual(msg.escape('^'), '\\S\\')
+        self.assertEqual(msg.escape('&'), '\\T\\')
+        self.assertEqual(msg.escape('~'), '\\R\\')
+
+        self.assertEqual(msg.escape('áéíóú'), '\\Xe1\\\\Xe9\\\\Xed\\\\Xf3\\\\Xfa\\')
+
+    def test_file(self):
+        # Extract message from file
+        self.assertTrue(hl7.isfile(sample_file))
+        messages = hl7.split_file(sample_file)
+        self.assertEqual(len(messages), 1)
+
+        # message can be parsed
+        msg = hl7.parse(messages[0])
+
+        # message has expected content
+        self.assertEqual([s[0][0] for s in msg], ['MSH', 'EVN', 'PID', 'PD1', 'NK1', 'PV1'])
 
 
 class IsHL7Test(unittest.TestCase):
@@ -77,6 +170,10 @@ class IsHL7Test(unittest.TestCase):
     def test_ishl7_wrongsegment(self):
         message = 'OBX|1|SN|1554-5^GLUCOSE^POST 12H CFST:MCNC:PT:SER/PLAS:QN||^182|mg/dl|70_105|H|||F\r'
         self.assertFalse(hl7.ishl7(message))
+
+    def test_isfile(self):
+        self.assertFalse(hl7.ishl7(sample_file))
+        self.assertTrue(hl7.isfile(sample_file))
 
 
 class ContainerTest(unittest.TestCase):
@@ -127,8 +224,8 @@ class ParsePlanTest(unittest.TestCase):
     def test_create_parse_plan(self):
         plan = hl7.create_parse_plan(sample_hl7)
 
-        self.assertEqual(plan.separators, ['\r', '|', '^'])
-        self.assertEqual(plan.containers, [Message, Segment, Field])
+        self.assertEqual(plan.separators, ['\r', '|', '~', '^', '&'])
+        self.assertEqual(plan.containers, [Message, Segment, Field, Repetition, Component])
 
     def test_parse_plan(self):
         plan = hl7.create_parse_plan(sample_hl7)
@@ -143,15 +240,23 @@ class ParsePlanTest(unittest.TestCase):
         plan = hl7.create_parse_plan(sample_hl7)
 
         n1 = plan.next()
-        self.assertEqual(n1.separators, ['|', '^'])
-        self.assertEqual(n1.containers, [Segment, Field])
+        self.assertEqual(n1.separators, ['|', '~', '^', '&'])
+        self.assertEqual(n1.containers, [Segment, Field, Repetition, Component])
 
         n2 = n1.next()
-        self.assertEqual(n2.separators, ['^'])
-        self.assertEqual(n2.containers, [Field])
+        self.assertEqual(n2.separators, ['~', '^', '&'])
+        self.assertEqual(n2.containers, [Field, Repetition, Component])
 
         n3 = n2.next()
-        self.assertTrue(n3 is None)
+        self.assertEqual(n3.separators, ['^', '&'])
+        self.assertEqual(n3.containers, [Repetition, Component])
+
+        n4 = n3.next()
+        self.assertEqual(n4.separators, ['&'])
+        self.assertEqual(n4.containers, [Component])
+
+        n5 = n4.next()
+        self.assertTrue(n5 is None)
 
 
 if __name__ == '__main__':